refactor(types): share snapshot manifest schema
