+++ crates/image/Cargo.toml
@@ -17,6 +17,7 @@ chrono.workspace = true
 futures.workspace = true
 hex.workspace = true
 libc.workspace = true
+microsandbox-types.workspace = true
 microsandbox-utils.workspace = true
 oci-client.workspace = true
 oci-spec.workspace = true
