+++ packages/microsandbox-types/rust/Cargo.toml
@@ -34,6 +34,7 @@ ts = [
 [dependencies]
 chrono.workspace = true
 dprint-plugin-typescript = { version = "=0.95", optional = true }
+hex.workspace = true
 ipnetwork.workspace = true
 microsandbox-types-macros.workspace = true
 serde.workspace = true
